+++ pkg/analyze/sequential.go
@@ -211,7 +211,7 @@ func (a *SequentialAnalyzer) processDir(path string) *Dir {
 
 	a.progressChan <- common.CurrentProgress{
 		CurrentItemName: path,
-		ItemCount:       len(files),
+		ItemCount:       int64(len(files)),
 		TotalSize:       totalSize,
 	}
 	return dir
