+++ pkg/analyze/sequential_coverage_test.go
@@ -33,7 +33,7 @@ func TestSequentialAnalyzerUpdateProgress(t *testing.T) {
 	// Send some progress updates
 	analyzer.progressChan <- struct {
 		CurrentItemName string
-		ItemCount       int
+		ItemCount       int64
 		TotalSize       int64
 	}{
 		CurrentItemName: "test",
@@ -60,7 +60,7 @@ func TestSequentialAnalyzerUpdateProgressWithDefaultCase(t *testing.T) {
 	// Send some progress updates
 	analyzer.progressChan <- struct {
 		CurrentItemName string
-		ItemCount       int
+		ItemCount       int64
 		TotalSize       int64
 	}{
 		CurrentItemName: "test",
@@ -74,7 +74,7 @@ func TestSequentialAnalyzerUpdateProgressWithDefaultCase(t *testing.T) {
 	// Send another progress update to trigger the default case
 	analyzer.progressChan <- struct {
 		CurrentItemName string
-		ItemCount       int
+		ItemCount       int64
 		TotalSize       int64
 	}{
 		CurrentItemName: "test2",
@@ -106,5 +106,5 @@ func TestSequentialAnalyzerAnalyzeDirWithIgnoreDir(t *testing.T) {
 	assert.NotNil(t, dir)
 	assert.Equal(t, "test_dir", dir.Name)
 	// Should have fewer items since nested directory was ignored
-	assert.Less(t, dir.ItemCount, 5)
+	assert.Less(t, dir.ItemCount, int64(5))
 }
